QQmlMetaObject::methodParameterTypes: Get metatype directly
In Qt 6, we can rely on QMetaMethod::metaType. Moreover, we can skip
searching for enumerators. In that case the metatype is necessarily
available (as enums cannot be forward declared, the type therefore
complete, and moc can generate the metatype already at compile time).

The last part also allows us to get rid of the isNamedEnumerator helper
methods.

Task-number: QTBUG-82931
Change-Id: Ic0c40a8fca8d6303db31e48bf6fb0169fd800f40
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>