CMake: Handle the case of QML resources in the BINARY directory
Qt Virtual Keyboard copies some QML resources to the BINARY directory
for dynamic configuration of resources.

_qt_internal_quick_compiler_process_resources assumes that all QML
source files are located under the SOURCE directory. This is perfectly
fine as long as the SOURCE and BINARY directories are located on the
same drive in Windows. Otherwise, the relative path cannot be resolved
and cmake is aborted because a directory cannot be created below.

Fix this problem by making sure that the relative path is resolved
against the correct base directory.

Pick-to: 6.1
Change-Id: I93e31e9f2720d448a48b6e8075cee0c01f59b584
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>