Introduce XmlListModel to QtDeclarative
XmlListModel was previously a part of QtXmlPatterns, which would not
be a part of Qt 6. The idea of this commit is to move a simplified
version of XmlListModel to QtDeclarative, so that it could be used
at least in the examples of different Qt modules.

Unlike the old implementation, this version does not have an XPath
support. This results in a reduced feature set - the user can't
build complicated XPath queries to populate model roles.
Now the user can select an xml element and, optionally, an
attribute, which will be used to extract the data.

[ChangeLog][XmlListModel] Introduce an XmlListModel QML model to
create read-only models from XML data.
This is a simplified version of a model from QtXmlPatterns, which
would no longer be a part of Qt 6. This model supports only simple
slash-separated paths and, optionally, one attribute for each
element.

Task-number: QTBUG-89817
Change-Id: I4186587dc1445dd981ac92b4ce104434236a32b9
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>