Add assertions and checks on QQWinPriv::deliveryAgentPrivate()
The static analyzer found the case where it could actually be null
during delivery of a WindowDeactivate event.  In other cases, we assert
rather than testing, until we know that it can actually be null.  It
shouldn't often be null because it's created in QQuickWindowPrivate::init()
and destroyed in the destructor; however we've already seen races when
event delivery continues while the window is being destroyed.  If more
such cases materialize, we can replace asserts with if()'s or check
the new inDestructor flag.

Fixes static analyzer warning 16738055c4c458145b59ad9b6bb643b5

Pick-to: 6.1
Change-Id: I408e9a7d71a77fd29c2af91e59caf9ede6585af9
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>