Add QQuickItemGrabResult::saveToFile(QUrl) overload
FileDialog provdes an url; alternatively something like
StandardPaths.writableLocation(StandardPaths.PicturesLocation) could
provide a good folder in which to save a screenshot file: also in the
form of a QUrl, but in that case, one might want to append a generated
filename.  Appending converts the URL to a string, and QUrl has no
QML-facing API to help modify it.  So depending on how it's done,
saveToFile() could be called with a plain filename, a path,
file:///path/to/file.png as a string or as a QUrl; it should now be able
to deal with all of those.

Also monospaced the true/false constants in qdoc comments.

Fixes: QTBUG-56436
Change-Id: I6058a52d3c64e243e44d4487f2b35fa31b1636fe
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>