CMake: Automatically add prefer directive to qmldir files
In qt6_add_qml_module we know that we are going to put all the QML files
listed in the generated qmldir into the resource path. We can thus
safely point the QML engine there.

This reveals that QtTest's plugin is actually not optional. It does
contain the compiled QML code.

Change-Id: Ia20f98c481482a851558c7ebeb39c9c365355d0a
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>