Simplify QQmlPropertyData::flagsForProperty; remove typeCategory
We do not need to call QQmlMetaType::typeCategory. The information
whether a type is a QObject or a QML list type is already stored in the
metaobject, and checked in the earlier if branches. Thus, typeCategory
could only ever return Unknown.

As this removes the only real caller of QQmlMetaType::typeCategory, we
can remove the function. QQmlEnginePrivate::typeCategory also called it,
but wasn't called itself (and is now also removed) .

Change-Id: Iea6055ba44de4d71334f40895e7ce85c3230f3d9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>