QQuickItem: reject hover events if not handled
This is the pattern we follow elsewhere in Qt - the event starts
out as accepted, and stays that way if it ends up being delivered
to an overridden event handler.  Otherwise the default handler
will reject it.
This change doesn't affect any of the pointerhandler auto tests.

This change is mostly a preparation for a later patch, which
relies on this standard behavior.

Change-Id: I158c828e958b5c1943ecd552a9594d24a3aaa90b
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>