QQmlAnyBinding: Honor interceptors if and only if requested
When installing a binding on a property which is intercepted, we can
either ignore or honor the interceptor (depending on whether the newly
added binding's value should be intercepted or not).
This was previously only implemented for QmlAbstractBinding based
bindings. This patch adds support for QProperty based bindings, and a
corresponding test.

Change-Id: Ib0a63426aa4a137cb7a375405843f2b0a97845d0
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>