Fix crash in QQmlPropertyBinding is not bound to property
In contrast to other QPropertyBindings, QQmlPropertyBinding could be
executed even if it is not installed on a property. This would happen if
we keep a strong reference to it, remove it from a property, and change
one of the observed JS dependencies. That would lead to a call of
expressionChanged, and then subsequently to an evaluate call. As setting
and removing the binding from a property can be done purely in C++, we
cannot detect this in the engine (which would allow us to disable the
binding, so that expressionChanged does not get called). Thus, we simply
check if the property data pointer is set when expressionChanged is
called, and return otherwise.

Task-number: QTBUG-89505
Change-Id: I933c1b3285d6472b1d2c6512ad37911090861298
Reviewed-by: Maximilian Goldstein <max.goldstein@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>