QQuickTableView: add API to get column widths and row heights
Add a set of functions that can be used to query both the actual
row and column sizes, but also what the implicit sizes are.
The implicit size of a column is defined as the maximum
implicit width found among the items in that column.
This implicit size is just a recommendation that can be used
by e.g HeaderView to resize a column to perfectly fit
the contents.

[ChangeLog][QtQuick][TableView] Added API to query row heights
and column widths: columnWidth(col), rowHeight(row),
implicitColumnWidth(col), implicitRowHeight(row).

Fixes: QTBUG-92124
Change-Id: Id8adbd558dab670d4d1c0bb268105b56b898b72a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>