QQuickFlickable: Avoid casting and rounding
The reportedVelocitySmoothing variable is never changed, and only used
as a parameter that is an int. So declare it as a const int instead of
qreal, and pass a floating-type zero for the destination parameter that
is of type qreal.

Fixes static analzyer warning 8e9457b62b3d1ff06f55ca46ba0e70ef

Change-Id: Ie9226a107dfaaf937d126833f8ba2786ac3d1e6b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>