qmllint: Only exclude default QML import dirs when explicitly requested
Previously specifying any additional QML import directories would lead to the default directories to be excluded.
This can be useful if you want to run qmllint on a project using a different Qt version but
this is not behavior you want in most cases.

[ChangeLog][qmllint][Important Behavior Change] Using -I no longer excludes default QML import directories.
Use --bare to get the old behavior.

Change-Id: I2ca5789cd44306bcad54f7b1b82049c21f721d3d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>