Consider only default constructed SourceLocation invalid
SourceLocation::isValid consider invalid all SourceLocations with
length == 0.
This is misleading because we already have several SourceLocations
that contain correct information but have zero length like:
* In the Ast several places, especially related to the use of
Semicolon in the grammar, which might be an automatic semicolon and
have length 0, but still be at the correct place, see ReturnStatement
and many others.
* Parser error messages, which (even worse) set only line/column

Note that SourceLocation() == SourceLocation(0,0,0,0) is indeed
invalid, as the first location of a file is SourceLocation(0,0,1,1)
because both column and line number start at 1.

Change-Id: Ib9184f2ee4842e85c505c94ccbbcefb119925492
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>