Prevent CMP0116 warning with CMake 3.20 or later
CMake 3.20 introduced a change in the handling of relative paths in *.d
depfiles referenced by custom commands. To avoid a CMP0116 policy
warning, we have to explicitly set the policy to NEW and also change
the relative path(s) we embed in the depfiles to be relative to
CMAKE_CURRENT_BINARY_DIR rather than the top level binary directory.

Fixes: QTBUG-92026
Pick-to: 6.1 6.0
Change-Id: I1a84d29f1a8d5c48bae5bc11596806f1e0e07919
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>